requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
coverage==7.8.2
Flask==3.1.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
pytest==8.3.5
pytest-cov==6.1.1
python-dotenv==1.1.0
pytz==2025.2
Werkzeug==3.1.3
